requirements.txt
chardet==5.2.0
Django==5.0.6
djangorestframework==3.15.2
google-cloud-storage==2.17.0
langchain_community==0.2.6
langchain_core==0.2.10
langchain_openai==0.1.10
langchain_text_splitters==0.2.2
langchain==0.2.6
pgvector==0.3.0
protobuf==5.27.2
psycopg2-binary==2.9.9
python-decouple==3.8